requirements.txt
# MASFIN System - Research Environment
# Compatible with Python 3.10+

python-dotenv>=1.0.0
crewai>=1.0.0
crewai-tools>=0.40.0
langchain>=1.0.0
langchain-community>=0.4.0
litellm>=1.70.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0
matplotlib>=3.7.0
prophet>=1.1.0
tenacity>=8.0.0
requests>=2.31.0
cryptography>=41.0.0
pyOpenSSL>=23.0.0
chroma-hnswlib>=0.7.0
jupyter>=1.0.0

# ── OmniSignal Additions ────────────────────────────────────────
fredapi>=0.5.2
pydantic>=2.6.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
aiohttp>=3.9.0
fpdf2>=2.7.0

# ── API Server ──────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ── Testing ─────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
